@@ -2,6 +2,7 @@
   "name": "knowit-kodekalender-frontend",
   "version": "1.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "serve": "vite",
     "preview": "vite preview --open --port 8080",
@@ -11,55 +12,53 @@
   },
   "dependencies": {
     "@headlessui/react": "^1.4.0",
-    "axios": "^0.21.0",
-    "clsx": "^1.1.1",
+    "axios": "^1.6.2",
+    "clsx": "^2.0.0",
     "date-fns": "^2.23.0",
-    "framer-motion": "^4.1.17",
+    "framer-motion": "^10.16.5",
     "lodash": "^4.17.20",
-    "react": "^18.0.0-alpha-8723e772b-20210826",
+    "react": "^18.2.0",
     "react-autosize-textarea": "^7.1.0",
-    "react-dom": "^18.0.0-alpha-8723e772b-20210826",
+    "react-dom": "^18.2.0",
     "react-hook-form": "^7.18.0",
     "react-icons": "^4.2.0",
     "react-query": "^3.23.2",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^6.19.0",
     "spark-md5": "^3.0.2",
-    "use-debounce": "^7.0.0"
+    "use-debounce": "^10.0.0"
   },
   "devDependencies": {
-    "@svgr/rollup": "^6.5.1",
-    "@tailwindcss/forms": "^0.3.4",
-    "@tailwindcss/line-clamp": "^0.2.1",
-    "@tailwindcss/typography": "^0.3.1",
-    "@types/jest": "^24.9.1",
+    "@optimize-lodash/rollup-plugin": "^4.0.4",
+    "@svgr/rollup": "^8.1.0",
+    "@tailwindcss/forms": "^0.5.7",
+    "@tailwindcss/typography": "^0.5.10",
+    "@types/jest": "^29.5.8",
     "@types/lodash": "^4.14.165",
-    "@types/node": "^12.12.62",
-    "@types/react": "^16.14.14",
-    "@types/react-dom": "^18.0.0",
+    "@types/node": "^20.9.2",
+    "@types/react": "^18.2.37",
+    "@types/react-dom": "^18.2.15",
     "@types/react-router-dom": "^5.1.6",
     "@types/spark-md5": "^3.0.2",
-    "@typescript-eslint/eslint-plugin": "^4.29.3",
-    "@typescript-eslint/parser": "^4.29.3",
-    "@vitejs/plugin-react-refresh": "^1.3.6",
+    "@typescript-eslint/eslint-plugin": "^6.11.0",
+    "@typescript-eslint/parser": "^6.11.0",
+    "@vitejs/plugin-react": "^4.2.0",
     "@whiterussianstudio/tailwind-easing": "^1.0.0",
     "autoprefixer": "^10.3.1",
-    "cssnano": "^5.0.8",
-    "eslint": "^7.32.0",
-    "eslint-plugin-import": "^2.24.2",
+    "cssnano": "^6.0.1",
+    "eslint": "^8.54.0",
+    "eslint-plugin-import": "^2.29.0",
     "eslint-plugin-lodash": "^7.3.0",
     "eslint-plugin-react": "^7.25.0",
     "eslint-plugin-react-hooks": "^4.2.0",
     "postcss": "^8.3.6",
     "postcss-100vh-fix": "^1.0.2",
-    "postcss-cli": "^8.3.1",
-    "rollup-plugin-optimize-lodash-imports": "^1.0.2",
+    "postcss-cli": "^10.1.0",
     "rollup-plugin-visualizer": "^5.5.2",
     "sass": "^1.38.2",
-    "svg-url-loader": "^6.0.0",
-    "tailwindcss": "^2.2.7",
+    "tailwindcss": "^3.3.5",
     "tailwindcss-children": "^2.1.0",
-    "typescript": "^4.4.0",
-    "vite": "^2.5.1"
+    "typescript": "^5.2.2",
+    "vite": "^5.0.0"
   },
   "browserslist": {
     "production": [
