@@ -1,6 +1,5 @@
 import { lazy, memo, Suspense, useCallback, useState } from "react"
-import { Switch, Route, Redirect } from "react-router-dom"
-import { map, range } from "lodash"
+import { Routes, Route, Navigate } from "react-router-dom"
 import { FaCogs } from "react-icons/fa"
 import clsx from "clsx"
 
@@ -32,8 +31,7 @@ const LazyAdmin = () => {
           "h-32",
           "translate-x-[-50%]",
           "translate-y-[-50%]",
-          "text-lightbulb-yellow",
-          "text-opacity-70",
+          "text-lightbulb-yellow/70",
           "animate-pulse"
         )}
       />
@@ -62,9 +60,6 @@ const App = () => {
 
   const [leaderboardHidden, setLeaderboardHidden] = useState(true)
 
-  // Match door 1-24 only
-  const doorPaths = map(range(1, 25), (door) => `/luke/:door(${door})`)
-
   const [bgAnimationPaused, setBgAnimationPaused] = useState(localStorage.getItem("stars-paused") === "true")
   const togglePaused = useCallback(() => {
     setBgAnimationPaused((state) => {
@@ -83,22 +78,20 @@ const App = () => {
     <div id="content-container" className="relative min-h-[calc(100vh+1.5rem)] h-[calc(100%+1.5rem)]">
       <Header className="-mb-4" setLeaderboardHidden={setLeaderboardHidden} />
 
-      <Switch>
-        <Route exact path="/" component={Doors} />
-        <Route path={doorPaths} component={Door} />
-        <Route path="/leaderboard" component={Leaderboard} />
-        <Route path="/gdpr" component={Gdpr} />
-        <Route path="/service_messages" component={ServiceMessages} />
-        <Route path="/solutions" component={Solutions} />
+      <Routes>
+        <Route path="/" element={<Doors />} />
+        <Route path="/luke/:door" element={<Door />} />
+        <Route path="/leaderboard" element={<Leaderboard />} />
+        <Route path="/gdpr" element={<Gdpr />} />
+        <Route path="/service_messages" element={<ServiceMessages />} />
+        <Route path="/solutions" element={<Solutions />} />
 
-        <Route path="/admin" component={LazyAdmin} />
-        <Route path="/users" component={LazyUser} />
+        <Route path="/admin" element={<LazyAdmin />} />
+        <Route path="/users" element={<LazyUser />} />
 
         {/* 404? - Route to main view */}
-        <Route>
-          <Redirect to="/" />
-        </Route>
-      </Switch>
+        <Route element={<Navigate to="/" />} />
+      </Routes>
 
       <BackgroundPauseButton paused={bgAnimationPaused} onTogglePaused={togglePaused} />
     </div>
