@@ -1,5 +1,5 @@
 import { useContext } from "react"
-import { useHistory } from "react-router"
+import { useNavigate } from "react-router-dom"
 
 import { useSignOut } from "../api/users/requests"
 import { AuthContext } from "../AuthContext"
@@ -8,7 +8,7 @@ import Button from "./Button"
 
 
 const SignOutButton = () => {
-  const history = useHistory()
+  const navigate = useNavigate()
 
   const { isAuthenticated } = useContext(AuthContext)
   const { mutate: signOut } = useSignOut()
@@ -18,7 +18,7 @@ const SignOutButton = () => {
   return (
     <Button
       tabIndex={4}
-      onClick={() => signOut(null, { onSuccess: () => history.push("/") })}
+      onClick={() => signOut(null, { onSuccess: () => navigate("/") })}
       content="Logg ut"
     />
   )
