@@ -1,5 +1,5 @@
-import { FC } from "react"
-import { Redirect, Route, Switch, useHistory } from "react-router"
+import { FC, useLayoutEffect } from "react"
+import { Navigate, Route, Routes, useNavigate } from "react-router-dom"
 
 import { useWhoami } from "../api/users/requests"
 import AdminHeader from "../components/Admin/AdminHeader"
@@ -15,30 +15,35 @@ import Page from "./Page"
 
 // TODO: Deleted posts? User list? Ban users?
 const Admin: FC = () => {
-  const history = useHistory()
+  const navigate = useNavigate()
 
   const { data: whoami, isLoading } = useWhoami()
 
-  // Limit access to admin pages. User can still perform admin actions
-  // programmatically, but they will be rejected by backend.
+  useLayoutEffect(() => {
+    // Limit access to admin pages. User can still perform admin actions
+    // programmatically, but they will be rejected by backend.
+    if (!isLoading && (!whoami || !whoami.is_admin))
+      navigate("/")
+  }, [isLoading, whoami, navigate])
+
   if (!isLoading && (!whoami || !whoami.is_admin))
-    history.push("/")
+    return null
 
   return (
     <Page className="py-12 px-8 md:px-12 mx-4 md:mx-8 bg-gray-100 text-gray-700 rounded-md space-y-8">
       <AdminHeader />
 
-      <Switch>
-        <Route exact path="/admin/doors" component={Doors} />
-        <Route path="/admin/doors/new" component={NewDoor} />
-        <Route path="/admin/doors/:door/edit" component={EditDoor} />
+      <Routes>
+        <Route path="/admin/doors" element={<Doors />} />
+        <Route path="/admin/doors/new" element={<NewDoor />} />
+        <Route path="/admin/doors/:door/edit" element={<EditDoor />} />
 
-        <Route exact path="/admin/service_messages" component={ServiceMessages} />
-        <Route path="/admin/service_messages/new" component={NewServiceMessage} />
-        <Route path="/admin/service_messages/:uuid/edit" component={EditServiceMessage} />
+        <Route path="/admin/service_messages" element={<ServiceMessages />} />
+        <Route path="/admin/service_messages/new" element={<NewServiceMessage />} />
+        <Route path="/admin/service_messages/:uuid/edit" element={<EditServiceMessage />} />
 
-        <Redirect to="/admin/doors" />
-      </Switch>
+        <Route element={<Navigate to="/admin/doors" />} />
+      </Routes>
     </Page>
   )
 }
