@@ -1,28 +1,29 @@
 import { find } from "lodash"
-import { useEffect, VFC } from "react"
-import { useHistory, useParams } from "react-router-dom"
+import { FC, useEffect } from "react"
+import { useNavigate, useParams } from "react-router-dom"
 
 import { useUpdateServiceMessage } from "../../api/admin/requests"
 import { AdminServiceMessagePayload } from "../../api/admin/ServiceMessage"
 import { useServiceMessages } from "../../api/requests"
 import ServiceMessageForm from "../../components/Admin/ServiceMessageForm"
+import { guardPresent } from "../../utils"
 
 
-const EditServiceMessage: VFC = () => {
+const EditServiceMessage: FC = () => {
   const { uuid } = useParams<{ uuid: string }>()
-  const history = useHistory()
+  const navigate = useNavigate()
 
   const { data: serviceMessage, isLoading } = useServiceMessages({ select: (serviceMessages) => find(serviceMessages, { uuid }) })
   const { mutate: updateServiceMessage } = useUpdateServiceMessage()
 
   const submit = (serviceMessage: AdminServiceMessagePayload) => {
-    updateServiceMessage({ uuid, service_message: serviceMessage }, { onSuccess: () => history.push("/admin/service_messages") })
+    guardPresent(uuid, (uuid) => updateServiceMessage({ uuid, service_message: serviceMessage }, { onSuccess: () => navigate("/admin/service_messages") }))
   }
 
   useEffect(() => {
     if (!isLoading && !serviceMessage)
-      history.push("/admin/service_messages/new")
-  }, [isLoading, serviceMessage, history])
+      navigate("/admin/service_messages/new")
+  }, [isLoading, serviceMessage, navigate])
 
   if (isLoading || !serviceMessage) return null
 
