@@ -1,18 +1,18 @@
-import { VFC } from "react"
-import { useHistory } from "react-router-dom"
+import { FC } from "react"
+import { useNavigate } from "react-router-dom"
 
 import { useCreateServiceMessage } from "../../api/admin/requests"
 import { AdminServiceMessagePayload } from "../../api/admin/ServiceMessage"
 import ServiceMessageForm from "../../components/Admin/ServiceMessageForm"
 
 
-const NewServiceMessage: VFC = () => {
-  const history = useHistory()
+const NewServiceMessage: FC = () => {
+  const navigate = useNavigate()
 
   const { mutate: createServiceMessage } = useCreateServiceMessage()
 
   const submit = (serviceMessage: AdminServiceMessagePayload) => {
-    createServiceMessage({ service_message: serviceMessage }, { onSuccess: () => history.push("/admin/service_messages") })
+    createServiceMessage({ service_message: serviceMessage }, { onSuccess: () => navigate("/admin/service_messages") })
   }
 
   return (
@@ -21,7 +21,7 @@ const NewServiceMessage: VFC = () => {
         <span className="text-4xl font-semibold">Ny driftsmelding</span>
       </div>
 
-      <ServiceMessageForm newForm serviceMessage={{ content: "", resolution_content: null, resolved_at: null }} submit={submit} />
+      <ServiceMessageForm newForm serviceMessage={{ content: "", resolution_content: null, resolved_at: null, door: null }} submit={submit} />
     </div>
   )
 }
