@@ -1,6 +1,6 @@
 import { ServiceMessage } from "../ServiceMessage"
 
 
-export type AdminServiceMessagePayload = Omit<ServiceMessage, "uuid" | "created_at"> & {
+export type AdminServiceMessagePayload = Omit<ServiceMessage, "uuid" | "created_at" | "resolved"> & {
   created_at?: string
 }
