@@ -1,5 +1,5 @@
 import { compact, map, sortBy, values } from "lodash"
-import { VFC } from "react"
+import { FC } from "react"
 
 import { useChallenges } from "../../api/admin/requests"
 
@@ -9,7 +9,7 @@ type DoorSelectProps = {
   setDoor: (door: number) => void
 }
 
-const DoorSelect: VFC<DoorSelectProps> = ({ door, setDoor }) => {
+const DoorSelect: FC<DoorSelectProps> = ({ door, setDoor }) => {
   const { data: challenges } = useChallenges()
 
   if (!challenges) return null
