@@ -16,7 +16,7 @@ const FormElement = forwardRef<HTMLInputElement, PropsWithChildren<FormElementPr
     <FormElementCustom label={label} note={note} disabled={disabled} className={labelClassName}>
       <input
         ref={ref}
-        className={clsx("block form-input", disabled && "border-opacity-30 border-gray-500", className)}
+        className={clsx("block form-input", disabled && "border-gray-500/30", className)}
         disabled={disabled}
         {...inputProps}
       >
